requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
regex
requests
pydantic
telebot
psycopg2
psycopg2-binary
